Make all() return a `bare` range
Changed my mind since the last commit :) . I don't intend for
enum_range to remain compatible with STL containers, as the view
concepts from Ranges STL likely to be standardized in the near future
are a better match for the intended use.

typedef `iterator` is left to ensure compatibility with Boost ranges.